Default to using an ephemeral port to serve the UI

+++ tests/functional/exporters/testing/test_webserver.py
@@ -2,6 +2,8 @@ import subprocess
 import time
 import sys
 import os
+import re
+import threading
 
 import pytest
 import requests
@@ -9,8 +11,11 @@ import requests
 from testplan.common.utils.process import kill_process
 from testplan import defaults
 
-TIMEOUT = 60
-REQUEST_TIMEOUT = 0.1
+_TIMEOUT = 60
+_REQUEST_TIMEOUT = 0.1
+_URL_RE = re.compile(
+    r'^View the JSON report in the browser: (?P<url>[^\s]+)\s*$')
+
 
 @pytest.yield_fixture(
     scope='module',
@@ -27,9 +32,13 @@ def dummy_testplan(request):
     """
     cmd = [sys.executable] + request.param
     cwd = os.path.dirname(os.path.abspath(__file__))
-    testplan_proc = subprocess.Popen(cmd, cwd=cwd)
+    testplan_proc = subprocess.Popen(cmd, cwd=cwd, stdout=subprocess.PIPE)
+
     yield testplan_proc
-    kill_process(testplan_proc)
+
+    if testplan_proc.poll() is None:
+        kill_process(testplan_proc)
+    assert testplan_proc.poll() is not None
 
 def test_webserver_exporter(dummy_testplan):
     """
@@ -37,19 +46,32 @@ def test_webserver_exporter(dummy_testplan):
     Repeatedly send requests to the web server until it answers or timeout is
     hit.
     """
-    url = 'http://{host}:{port}/testplan/local'.format(
-        host=defaults.WEB_SERVER_HOSTNAME,
-        port=defaults.WEB_SERVER_PORT
-    )
-    timeout = time.time() + TIMEOUT
+    # By default Testplan will grab an ephemeral port to serve the UI, so we
+    # must parse the stdout to find the URL.
+    url = None
+
+    while (url is None) and (dummy_testplan.poll() is None):
+        stdout_line = dummy_testplan.stdout.readline().decode('ascii')
+        match = _URL_RE.match(stdout_line)
+        if match:
+            url = match.group('url')
+
+    assert url is not None, 'Failed to parse the webserver URL'
+
+    # Now that we have the URL, try to make a GET request to it. This might
+    # not immediately succeed so try a few times allowing for connection
+    # errors. When the GET response is received, just verify the status code
+    # is 200 OK.
+    timeout = time.time() + _TIMEOUT
     status_code = None
     while time.time() < timeout:
         try:
-            response = requests.get(url, timeout=REQUEST_TIMEOUT)
+            response = requests.get(url, timeout=_REQUEST_TIMEOUT)
         except (requests.exceptions.ConnectionError,
                 requests.exceptions.ReadTimeout):
-            time.sleep(REQUEST_TIMEOUT)
+            time.sleep(_REQUEST_TIMEOUT)
         else:
             status_code = response.status_code
             break
     assert status_code == 200
+
